src/main/java/com/crud/tasks/trello/client/TrelloClient.java
package com.crud.tasks.trello.client;

import com.crud.tasks.domain.CreatedTrelloCardDto;
import com.crud.tasks.domain.TrelloBoardDto;
import com.crud.tasks.domain.TrelloCardDto;
import com.crud.tasks.trello.config.TrelloConfig;
import lombok.RequiredArgsConstructor;
import org.slf4j.Logger;
import org.slf4j.LoggerFactory;
import org.springframework.stereotype.Component;
import org.springframework.web.client.RestClientException;
import org.springframework.web.client.RestTemplate;
import org.springframework.web.util.UriComponentsBuilder;

import java.net.URI;
import java.util.Arrays;
import java.util.Collections;
import java.util.List;
import java.util.Optional;
import java.util.stream.Collectors;

@Component
@RequiredArgsConstructor
public class TrelloClient {

    private static final Logger LOGGER = LoggerFactory.getLogger(TrelloClient.class);

    private final RestTemplate restTemplate;
    private final TrelloConfig trelloConfig;


    public URI buildTrelloUrlGetBoards () {
        URI url = UriComponentsBuilder.fromHttpUrl(trelloConfig.getTrelloApiEndpoint() + "/members/" + trelloConfig.getUserName() + "/boards")
                .queryParam("key", trelloConfig.getTrelloAppKey())
                .queryParam("token", trelloConfig.getTrelloToken())
                .queryParam("fields", "id,name")
                .queryParam("lists", "all")
                .build()
                .encode()
                .toUri();
        return url;
    }

    public URI buildTrelloUrlCardPost (TrelloCardDto trelloCardDto) {
        URI url = UriComponentsBuilder.fromHttpUrl(trelloConfig.getTrelloApiEndpoint() + "/cards")
                .queryParam("key", trelloConfig.getTrelloAppKey())
                .queryParam("token", trelloConfig.getTrelloToken())
                .queryParam("name", trelloCardDto.getName())
                .queryParam("desc", trelloCardDto.getDescription())
                .queryParam("pos", trelloCardDto.getPos())
                .queryParam("idList", trelloCardDto.getListId())
                .build()
                .encode()
                .toUri();
        return url;
    }

    public CreatedTrelloCardDto createNewCard (TrelloCardDto trelloCardDto) {
        return restTemplate.postForObject(buildTrelloUrlCardPost(trelloCardDto), null, CreatedTrelloCardDto.class);
    }

    public List<TrelloBoardDto> getTrelloBoards () {

        URI url = buildTrelloUrlGetBoards();
        try {
            TrelloBoardDto[] boardsResponse = restTemplate.getForObject(url,
                    TrelloBoardDto[].class);
            return Optional.ofNullable(boardsResponse)
                    .map(Arrays::asList)
                    .orElse(Collections.emptyList())
                    .stream()
                    .filter(n -> n.getId() != null && n.getName() != null)
                    .collect(Collectors.toList());
        } catch (RestClientException e) {
                LOGGER.error(e.getMessage(), e);
                return Collections.emptyList();
        }
    }
}
